#104 refactor: split videopicker and videofeed
