+++ src/VideoComponents/VideoApp.jsx
@@ -11,6 +11,7 @@ import MiniMapWidget from './MiniMapWidget';
 import GamepadWrapper from './GamepadWrapper';
 import KeyboardWrapper from './KeyboardWrapper';
 import MenuButton from './MenuButton';
+import VideoPicker from './VideoPicker';
 
 const { remote } = window.require('electron');
 
@@ -19,6 +20,7 @@ function VideoApp() {
   const [sensorValues, sensorUpdate] = useState(remote.getGlobal('fromROV'));
   const [biasValues, biasUpdate] = useState(remote.getGlobal('bias'));
   const [transparent, toggleTransparent] = useState(false);
+  const [deviceId, setDeviceId] = useState('');
 
   const closeVideoWindow = () => {
     let w = remote.getCurrentWindow();
@@ -40,6 +42,11 @@ function VideoApp() {
           clickFunction={() => toggleTransparent(!transparent)}
           image="transparent"
           additionalClass="transparentBtn"
+        />
+        <VideoPicker
+          deviceId={deviceId}
+          handleClick={id => setDeviceId(id)}
+          hidden={transparent}
         />
         <MenuButton
           clickFunction={() => {
@@ -77,7 +84,7 @@ function VideoApp() {
       />
       <GamepadWrapper className="GamepadWrapper" />
       <KeyboardWrapper className="KeyboardInput" />
-      <VideoFeed hidden={transparent} />
+      <VideoFeed deviceId={deviceId} hidden={transparent} />
     </div>
   );
 }
